Correct the bit-width of xemacs_c_alloca()'s SIZE argument.
src/ChangeLog addition:

2017-11-23  Aidan Kehoe  <kehoea@parhasard.net>

	* alloca.c (xemacs_c_alloca):
	If this is going to take an unsigned argument, as it always has,
	that unsigned argument needs to have the bit width of a
	size_t. Use it explicitly.
	Do some overflow checking, too, calling memory_full() if we find
	it.
	* lisp.h:
	Updatte the declaration of xemacs_c_alloca().

tests/ChangeLog addition:

2017-11-23  Aidan Kehoe  <kehoea@parhasard.net>

	* automated/format-tests.el (args-out-of-range):
	We can give out-of-memory errors too with
	(format (concat "%" (number-to-string most-positive-fixnum) "d"),
	and that's appropriate.

+++ src/lisp.h
@@ -1492,7 +1492,7 @@ we do two things:
   MULTIUSE_ALLOCA (size, sizeagain)
 #endif /* (not) BROKEN_ALLOCA_IN_FUNCTION_CALLS */
 
-MODULE_API void *xemacs_c_alloca (unsigned int size) ATTRIBUTE_MALLOC;
+MODULE_API void *xemacs_c_alloca (size_t size) ATTRIBUTE_MALLOC;
 
 MODULE_API int record_unwind_protect_freeing (void *ptr);
